requirements.txt
modal
requests
gspread
google-auth
python-dotenv
